+++ packages/google_maps_flutter/google_maps_flutter/lib/src/controller.dart
@@ -18,6 +18,14 @@ class GoogleMapController {
   /// The mapId for this controller
   final int mapId;
 
+  /// List of active stream subscriptions for map events.
+  ///
+  /// This list keeps track of all event subscriptions created for the map,
+  /// including camera movements, marker interactions, and other map events.
+  /// These subscriptions are properly disposed when the controller is disposed.
+  final List<StreamSubscription<dynamic>> _streamSubscriptionList =
+      <StreamSubscription<dynamic>>[];
+
   /// Initialize control of a [GoogleMap] with [id].
   ///
   /// Mainly for internal use when instantiating a [GoogleMapController] passed
@@ -38,53 +46,108 @@ class GoogleMapController {
 
   void _connectStreams(int mapId) {
     if (_googleMapState.widget.onCameraMoveStarted != null) {
-      GoogleMapsFlutterPlatform.instance
-          .onCameraMoveStarted(mapId: mapId)
-          .listen((_) => _googleMapState.widget.onCameraMoveStarted!());
+      _addSubscription(
+        GoogleMapsFlutterPlatform.instance
+            .onCameraMoveStarted(mapId: mapId)
+            .listen((_) => _googleMapState.widget.onCameraMoveStarted!()),
+      );
     }
     if (_googleMapState.widget.onCameraMove != null) {
-      GoogleMapsFlutterPlatform.instance.onCameraMove(mapId: mapId).listen(
-          (CameraMoveEvent e) => _googleMapState.widget.onCameraMove!(e.value));
+      _addSubscription(
+        GoogleMapsFlutterPlatform.instance.onCameraMove(mapId: mapId).listen(
+              (CameraMoveEvent e) =>
+                  _googleMapState.widget.onCameraMove!(e.value),
+            ),
+      );
     }
     if (_googleMapState.widget.onCameraIdle != null) {
-      GoogleMapsFlutterPlatform.instance
-          .onCameraIdle(mapId: mapId)
-          .listen((_) => _googleMapState.widget.onCameraIdle!());
+      _addSubscription(
+        GoogleMapsFlutterPlatform.instance
+            .onCameraIdle(mapId: mapId)
+            .listen((_) => _googleMapState.widget.onCameraIdle!()),
+      );
     }
-    GoogleMapsFlutterPlatform.instance
-        .onMarkerTap(mapId: mapId)
-        .listen((MarkerTapEvent e) => _googleMapState.onMarkerTap(e.value));
-    GoogleMapsFlutterPlatform.instance.onMarkerDragStart(mapId: mapId).listen(
-        (MarkerDragStartEvent e) =>
-            _googleMapState.onMarkerDragStart(e.value, e.position));
-    GoogleMapsFlutterPlatform.instance.onMarkerDrag(mapId: mapId).listen(
-        (MarkerDragEvent e) =>
-            _googleMapState.onMarkerDrag(e.value, e.position));
-    GoogleMapsFlutterPlatform.instance.onMarkerDragEnd(mapId: mapId).listen(
-        (MarkerDragEndEvent e) =>
-            _googleMapState.onMarkerDragEnd(e.value, e.position));
-    GoogleMapsFlutterPlatform.instance.onInfoWindowTap(mapId: mapId).listen(
-        (InfoWindowTapEvent e) => _googleMapState.onInfoWindowTap(e.value));
-    GoogleMapsFlutterPlatform.instance
-        .onPolylineTap(mapId: mapId)
-        .listen((PolylineTapEvent e) => _googleMapState.onPolylineTap(e.value));
-    GoogleMapsFlutterPlatform.instance
-        .onPolygonTap(mapId: mapId)
-        .listen((PolygonTapEvent e) => _googleMapState.onPolygonTap(e.value));
-    GoogleMapsFlutterPlatform.instance
-        .onCircleTap(mapId: mapId)
-        .listen((CircleTapEvent e) => _googleMapState.onCircleTap(e.value));
-    GoogleMapsFlutterPlatform.instance.onGroundOverlayTap(mapId: mapId).listen(
-        (GroundOverlayTapEvent e) =>
-            _googleMapState.onGroundOverlayTap(e.value));
-    GoogleMapsFlutterPlatform.instance
-        .onTap(mapId: mapId)
-        .listen((MapTapEvent e) => _googleMapState.onTap(e.position));
-    GoogleMapsFlutterPlatform.instance.onLongPress(mapId: mapId).listen(
-        (MapLongPressEvent e) => _googleMapState.onLongPress(e.position));
-    GoogleMapsFlutterPlatform.instance
-        .onClusterTap(mapId: mapId)
-        .listen((ClusterTapEvent e) => _googleMapState.onClusterTap(e.value));
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance
+          .onMarkerTap(mapId: mapId)
+          .listen((MarkerTapEvent e) => _googleMapState.onMarkerTap(e.value)),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance.onMarkerDragStart(mapId: mapId).listen(
+            (MarkerDragStartEvent e) =>
+                _googleMapState.onMarkerDragStart(e.value, e.position),
+          ),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance.onMarkerDrag(mapId: mapId).listen(
+            (MarkerDragEvent e) =>
+                _googleMapState.onMarkerDrag(e.value, e.position),
+          ),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance.onMarkerDragEnd(mapId: mapId).listen(
+            (MarkerDragEndEvent e) =>
+                _googleMapState.onMarkerDragEnd(e.value, e.position),
+          ),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance.onInfoWindowTap(mapId: mapId).listen(
+            (InfoWindowTapEvent e) => _googleMapState.onInfoWindowTap(e.value),
+          ),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance.onPolylineTap(mapId: mapId).listen(
+            (PolylineTapEvent e) => _googleMapState.onPolylineTap(e.value),
+          ),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance.onPolygonTap(mapId: mapId).listen(
+            (PolygonTapEvent e) => _googleMapState.onPolygonTap(e.value),
+          ),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance
+          .onCircleTap(mapId: mapId)
+          .listen((CircleTapEvent e) => _googleMapState.onCircleTap(e.value)),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance
+          .onTap(mapId: mapId)
+          .listen((MapTapEvent e) => _googleMapState.onTap(e.position)),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance.onLongPress(mapId: mapId).listen(
+            (MapLongPressEvent e) => _googleMapState.onLongPress(e.position),
+          ),
+    );
+    _addSubscription(
+      GoogleMapsFlutterPlatform.instance.onClusterTap(mapId: mapId).listen(
+            (ClusterTapEvent e) => _googleMapState.onClusterTap(e.value),
+          ),
+    );
+  }
+
+  /// Adds a stream subscription to the list of active subscriptions.
+  ///
+  /// This method is used to track and manage all stream subscriptions
+  /// created for map events. The subscriptions are stored in
+  /// [_streamSubscriptionList] and will be properly disposed when the
+  /// controller is disposed.
+  void _addSubscription(StreamSubscription<dynamic> subscription) {
+    _streamSubscriptionList.add(subscription);
+  }
+
+  /// Adds a stream subscription to the tracked list for testing purposes.
+  ///
+  /// This method allows tests to add custom stream subscriptions to the
+  /// controller's tracking list. It is only intended for use in tests and
+  /// should not be called in production code.
+  ///
+  /// The added subscription will be properly canceled when the controller
+  /// is disposed, allowing tests to verify subscription cleanup behavior.
+  @visibleForTesting
+  void addDebugTrackSubscription(StreamSubscription<dynamic> subscription) {
+    _addSubscription(subscription);
   }
 
   /// Updates configuration options of the map user interface.
@@ -321,6 +384,11 @@ class GoogleMapController {
 
   /// Disposes of the platform resources
   void dispose() {
+    for (final StreamSubscription<dynamic> streamSubscription
+        in _streamSubscriptionList) {
+      streamSubscription.cancel();
+    }
+    _streamSubscriptionList.clear();
     GoogleMapsFlutterPlatform.instance.dispose(mapId: mapId);
   }
 }