+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.12.2
+
+* Fixes memory leak by implementing mechanism for disposing stream subscriptions in `GoogleMapController`. 
+  Previously, subscriptions were not canceled upon disposal of the controller, which could lead to memory leaks or
+  unexpected behavior. The controller now tracks all stream subscriptions and cancels them when `dispose()` is called.
+
 ## 2.12.1
 
 * Fixes typo in README.